celery memory/connection leaks workaround
Celery is spawned after stevedore plugins are loaded

But at startup the subscription plugins load the entire database to
setup rabbitmq subscription, opening the connection to rabbitmq and
keeping in RAM all subscriptions callback configuration.

Then Celery main worker is spawned with multiprocessing.Process().  This
one uses fork() and then fork() is used again to spawn other worker() to
run tasks.

After fork() occurs no resources are freed, the FD to rabbitmq is still
open and previously created resources are still there but unused.

To avoid that this spawns the celery workers earlier.

In a second step, the celery workers should be setuped as Celery
recommemed (using "celery worker" command), so only the celery tasks
will be loaded in memory of celery workers and nothing else.