http callback: fix handling of body charset.
The RFC default charset is us-ascii, but if not provide we encode it in
UTF8. Everybody agree that UTF8 is better default now, but web server
must respect the RFC and use us-acsii if charset is not provided.

Infortunatly mockserver was not respecting the RFC and was decoding the
body as utf8 instead of us-ascii, making some test to pass instead of
failing.

This change ensure we always have a charset. If the subscription does
not have it we uses utf-8, if subscription have it we encode the body
with the provided one.

So servers will always known what the encoding is.