Remove logs relationship of Subscription orm object
This attribute is no more used.

If subscription orm object is loaded, due to the lazy=joined, the whole logs table
is loaded too and maybe refreshed along as we run hooks in certain case.

It's still unclear why the memory wasn't freed when this orm object is
loaded. It's maybe due to the manual construction of the sqlachemy
Mapper object.

A future change will remove the customer sqla mapper to the usual
declarative_base().